requirements.txt
anthropic[vertex]>=0.40.0
pydantic[email]>=2.0.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
httpx>=0.27.0
jinja2>=3.1.0
python-dateutil>=2.8.0
cryptography>=41.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
slowapi>=0.1.9
pdfplumber>=0.10.0
python-docx>=1.1.0
# PostgreSQL (Cloud SQL + Supabase)
psycopg2>=2.9.0
# GCP services
google-cloud-secret-manager>=2.18.0
google-cloud-aiplatform>=1.43.0
google-auth>=2.28.0
# Auth (Cloud Identity Platform)
PyJWT>=2.8.0
# Email
resend>=2.0.0
